# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TrulyBot API",
    "description": "AI-powered chatbot platform API for e-commerce businesses",
    "version": "1.0.0",
    "contact": {
      "name": "TrulyBot Support",
      "url": "https://trulybot.xyz/contact",
      "email": "support@trulybot.xyz"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://trulybot.xyz/api",
      "description": "Production server"
    },
    {
      "url": "https://staging.trulybot.xyz/api",
      "description": "Staging server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/payments/create-order": {
      "post": {
        "tags": ["Payments"],
        "summary": "Create payment order",
        "description": "Create a new payment order for subscription",
        "operationId": "createPaymentOrder",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrder"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment order created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentOrder"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT token"
      }
    },
    "schemas": {
...
      "CreateOrder": {
        "type": "object",
        "required": ["planId", "userId"],
        "properties": {
          "planId": {
            "type": "string",
            "enum": ["basic", "pro", "ultra"]
          },
          "currency": {
            "type": "string",
            "enum": ["INR", "USD"],
            "default": "INR"
          },
          "billingPeriod": {
            "type": "string",
            "enum": ["monthly", "yearly"],
            "default": "monthly"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "PaymentOrder": {
        "type": "object",
        "required": ["orderId", "amount", "currency"],
        "properties": {
          "orderId": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": ["success", "error"],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [false]
          },
          "error": {
            "type": "object",
            "required": ["message"],
            "properties": {
              "message": {
                "type": "string"
              },
              "code": {
                "type": "string"
              },
              "details": {
                "type": "object"
              }
            }
          },
          "meta": {
            "type": "object",
            "required": ["timestamp", "requestId"],
            "properties": {
              "timestamp": {
                "type": "string",
                "format": "date-time"
              },
              "requestId": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
...
}